Set all components initialized only after backup
We set /var/lib/all_components_initialized when all components are initialized.
That will be use to check and configure clusters (both ZKS and edge node cluster).
But we also take backup of single node config to /var/lib so that we can factory reset
and switch back to single node mode if user deletes cluster.

There is a timing issue here, we are prematutely setting the flag all_components_initialized before the
backup is saved, so in some race conditions we run into an issue where backup was not saved before switching to
cluster mode. This commit fixes that.

Signed-off-by: Pramodh Pallapothu <pramodh@zededa.com>
(cherry picked from commit 4621069290be408988e8db591c4f65e7d1a14184)